+++ app/src/main/java/com/example/mspaint/mainui/drawingScreen.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.StrokeCap
 import androidx.compose.ui.input.pointer.pointerInput
-import com.example.mspaint.canvasObjectData.Line
+import com.example.mspaint.attributes.Line
 
 @Composable
 fun DrawScreen() {
