+++ app/src/main/java/com/example/mspaint/tools/Eraser.kt
@@ -1,6 +1,6 @@
 package com.example.mspaint.tools
 
-import com.example.mspaint.canvasObjectData.hue
+import com.example.mspaint.attributes.hue
 
 fun eraser () {
     // TODO: change the hue to the color of the canvas which isn't white
