+++ app/src/main/java/com/example/mspaint/tools/Pencil.kt
@@ -1,6 +1,6 @@
 package com.example.mspaint.tools
 
-import com.example.mspaint.canvasObjectData.hue
+import com.example.mspaint.attributes.hue
 
 fun pencil() {
     hue = androidx.compose.ui.graphics.Color.Black
