--- app/src/main/java/com/example/mspaint/ui/theme/projectColorPalette.kt
@@ -1,28 +1,0 @@
-package com.example.mspaint.ui.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-val PureBlack = Color(0xFF000000)
-
-// color pallet
-val Red = Color(0xFFff0000)
-val Orange = Color(0xFFff8000)
-val Yellow = Color(0xFFffff00)
-val Green = Color(0xFF00ff00)
-val Aqua = Color(0xFF00ffff)
-val White = Color(0xFFffffff)
-val Black = Color(0xFF1a1a1a)
-val Purple = Color(0xFF8000ff)
-val Magenta = Color(0xFFff00ff)
-val Blue = Color(0xFF0000ff)
-
-// toolbar
-val Grey = Color(0xFFc8c8c8)
